Load and save totally work

+++ Shared/src/archeologyp2/shared/finds/Hearth.java
@@ -9,8 +9,8 @@ package archeologyp2.shared.finds;
  */
 public class Hearth extends Charcoal {
 
-	private int length;
-	private int width;
+	private double length;
+	private double width;
 	
 	/**
 	 * 
@@ -26,7 +26,7 @@ public class Hearth extends Charcoal {
 	 * @param length
 	 * @param width
 	 */
-	public Hearth(int date, int length, int width) {
+	public Hearth(int date, double length, double width) {
 		super(date);
 		this.length = length;
 		this.width = width;
@@ -36,7 +36,7 @@ public class Hearth extends Charcoal {
 	 * 
 	 * @return length
 	 */
-	public int getLength(){
+	public double getLength(){
 		return length;
 	}
 	
@@ -44,7 +44,7 @@ public class Hearth extends Charcoal {
 	 * 
 	 * @return width
 	 */
-	public int getWidth(){
+	public double getWidth(){
 		return width;
 	}
